+++ src/Components/UnkownGuessMsg/UnknownGuessMsg.tsx
@@ -1,0 +1,10 @@
+import './unknownGuess.css';
+
+export default function UnknownGuessMsg({ inputValue }: {inputValue: string} ) {
+    console.log(inputValue);
+    return (
+        <div className="unknown-guess-msg">
+            Neznana občina: { inputValue }
+        </div>
+    )
+}