+++ src/Components/WrongGuessMsg/WrongGuessMsg.tsx
@@ -2,7 +2,7 @@ import './wrongGuessMsg.css';
 
 export default function WrongGuessMsg() {
     return (
-        <div className="wrong-guess-message">
+        <div className="wrong-guess-msg">
             Wrong guess! Try again.
         </div>
     )