+++ src/types/input.ts
@@ -1,5 +1,3 @@
-import { Features } from "./game";
-
 export type InputEvent = React.ChangeEvent<HTMLInputElement>;
 export type FormEvent = React.FormEvent<HTMLFormElement>;
 export type ClickEvent = React.MouseEvent<HTMLElement>;
@@ -10,5 +8,5 @@ export interface InputProps {
     setInputValue: React.Dispatch<React.SetStateAction<string>>;
     handleGuess: (guess: string) => void;
     numberOfGuesses: number;
-    allFeatures: Features;
+    obcine: string[];
 }